// src/api/coreTypes.ts
export interface CategoryGetByIdType {
  id: number;
  name: string;
  description?: string | null;
}

export interface CategoryCreateType {
  name: string;
  description?: string | null;
}

export interface CategoryEditType {
  id: number;
  name?: string | null;
  description?: string | null;
}